+++ CSS/styles.css
@@ -1,0 +1,131 @@
+*{
+    padding: 0;
+    margin: 0;
+}
+body{
+    font-family: Arial, Helvetica, sans-serif;
+    font-size: 18px;
+    line-height: 2;
+}
+header{
+    background-image:url("../files/fond.png");
+    background-size: cover;
+    color: #fdf0d1;
+    padding: 15px 0;
+    display: flex;
+    justify-content: space-around;
+}
+header div{
+    text-align: center;
+    border-radius: 10px;
+    margin: auto 0;
+    border: 3px double #fdf0d1;
+    width: 200px;
+    height: 35px;
+}
+nav{
+    padding: 15px 0;
+    text-align: center;
+    background-color: #fdf0d1 ;
+}
+ul{
+    list-style: none;
+}
+ul li{
+    width: 200px;
+    display: inline-block;
+    background-color: #616d81;
+    margin: 0 15px ;
+    border-radius: 10px;
+}
+ul li:hover{
+    cursor: pointer;
+    transition: 0.5s;
+    background-color: #313d52;
+    color: #f8d07f;
+}
+ul li a{
+    text-decoration: none;
+    color: #fdf0d1;
+}
+main{
+    background-color: #fdf0d1 ;
+    height: 500px;
+    padding: 30px 0 0 0 ;
+}
+main section{
+    margin: auto;
+    height: 85%;
+    box-shadow: 1px 2px 3px 8px rgba(49, 12, 40,0.5);
+    border-radius: 15px;
+    width: 40%;
+}
+main section iframe{
+    width: 100%;
+    height: 100%;
+    border-radius: 15px;
+}
+footer{
+    text-align: center;
+    color: #fdf0d1;
+    padding: 23px;
+    background-color: #616d81;
+}
+fieldset{
+    border: 4px solid #616d81;
+    border-radius: 18px;
+    width: 95%;
+    margin: 10px auto ;
+    text-align: center;
+}
+fieldset legend{
+    padding: 15px;
+    color: #616d81;
+    font-size: 25px;
+    text-decoration: underline;
+}
+hr{
+    margin: 15px 0;
+    border: 2px solid #616d81;
+}
+fieldset input[type=text],fieldset input[type=number],fieldset select,fieldset input[type=reset],fieldset input[type=submit]{
+    padding: 7px;
+    border-radius: 15px;
+    border: 2px double #616d81;
+    margin: 5px 0;
+    font-size: 18px;
+    width: 250px;
+}
+fieldset input::placeholder{
+    text-align: center;
+}
+fieldset input[type=reset],fieldset input[type=submit]{
+    width: 150px;
+    font-weight: bolder;
+    color: #616d81;
+}
+fieldset input:hover{
+    cursor: pointer;
+    transition: 1s;
+    letter-spacing: 2px;
+    background-color: #616d81;
+    color: #fdf0d1;
+}
+fieldset input:hover::placeholder{
+    color: #fdf0d1;
+}
+table{
+    border-collapse: collapse;
+    text-align: center;
+}
+table th {
+    background-color: #616d81;
+}
+table td ,table th{
+    border: 3px solid #313d52;
+    padding: 7px 15px;
+    width: 200px;
+}
+h4{
+    margin: 10px 0;
+}
