+++ HTML/ajoutabonne.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+    <link rel="stylesheet" href="../CSS/styles.css" />
+  </head>
+  <body>
+    <form action="../PHP/ajoutabonne.php" method="get">
+      <fieldset>
+        <legend>Ajout d'un abonné</legend>
+        <input
+          type="text"
+          name="cin"
+          id="cin"
+          placeholder="CIN de l'abonné"
+        /><br />
+        <input
+          type="text"
+          name="nom"
+          id="nom"
+          placeholder="Nom de l'abonné"
+        /><br />
+        <input
+          type="text"
+          name="prenom"
+          id="prenom"
+          placeholder="Prenom de l'abonné"
+        /><br />
+        <input type="radio" name="sexe" id="H" value="H"><label for="H">Homme</label>
+        <input type="radio" name="sexe" id="F" value="F"><label for="F">Femme</label><br>
+        <input
+        type="text"
+        name="fonction"
+        id="fonction"
+        placeholder="Fonction de l'abonné"
+      />
+        <hr />
+        <input type="submit" value="Ajouter" />
+        <input type="reset" value="Annuler" />
+      </fieldset>
+    </form>
+  </body>
+</html>
