+++ HTML/ajoutlivre.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+    <link rel="stylesheet" href="../CSS/styles.css" />
+  </head>
+  <body>
+    <form action="../PHP/ajoutlivre.php" method="get">
+      <fieldset>
+        <legend>Ajout d'un livre</legend>
+        <input
+          type="text"
+          name="titre"
+          id="titre"
+          placeholder="Titre du livre"
+        /><br />
+        <input
+          type="text"
+          name="type"
+          id="type"
+          placeholder="Type du livre"
+        /><br />
+        <input
+          type="number"
+          name="nbr"
+          id="nbr"
+          placeholder="Nombre d'exemplaires"
+        /><br />
+        <hr />
+        <input type="submit" value="Ajouter" />
+        <input type="reset" value="Annuler" />
+      </fieldset>
+    </form>
+  </body>
+</html>
