+++ JS/date.js
@@ -1,0 +1,5 @@
+function affichedate() {
+  const now = new Date();
+  const date = now.toString();
+  document.getElementById("date").innerHTML = date.substr(0, 15);
+}
