+++ PHP/connection.php
@@ -1,0 +1,3 @@
+<?php
+$connect = mysqli_connect("localhost","root","","biblio");
+?>
