--- connection.php
@@ -1,3 +1,0 @@
-<?php
-$connect = mysqli_connect("localhost","root","","biblio");
-?>
